backend/suzie_api/requirements.txt
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
boto3==1.28.62
botocore==1.31.62
certifi==2023.7.22
charset-normalizer==3.3.0
Django==3.2
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
jmespath==1.0.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==9.5.0
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
replicate==0.15.3
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6